+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.4
+* Added an option to change gold and currency icon to the left side.
+* Added an option to display gold for both factions in the tooltip.
+* TOC bump for patch 7.1.0.
+
 ## 1.1.3
 * Addon now hooks directly to mail functions for improved compability with other addons.
 * Added option to reset history data from today and yesterday.